Fix guided onboarding analysis page exports
