+++ app/dashboard/onboarding-v2/[sessionId]/summary/page.tsx
@@ -3,6 +3,7 @@ import GuidedPageStepPanel from "@/features/onboarding-v2/components/GuidedPageS
 import RunAnalysisButton from "./RunAnalysisButton";
 import { requireAdminPageAccess } from "@/features/shared/lib/server/admin-access";
 import { createServerSupabaseRSC } from "@/features/shared/lib/supabase/server";
+import { filterGuidedAnalysisRecommendations } from "@/features/onboarding-v2/analysis/filterGuidedAnalysisRecommendations";
 import type { AiRecommendationRecord } from "@/features/ai/server/types";
 import type { Json } from "@shared/types/types/supabase";
 
@@ -48,29 +49,6 @@ function jsonHasContent(value: Json | null | undefined): boolean {
   if (Array.isArray(value)) return value.length > 0;
   if (typeof value === "object") return Object.keys(value).length > 0;
   return Boolean(value);
-}
-
-function metadataMatchesSession(metadata: Json | null | undefined, sessionId: string): boolean {
-  if (!metadata || typeof metadata !== "object" || Array.isArray(metadata)) return false;
-  return metadata.guidedSessionId === sessionId || metadata.sessionId === sessionId;
-}
-
-function isOnboardingScopedRecommendation(recommendation: AiRecommendationRecord, sessionId: string): boolean {
-  return recommendation.domain === "onboarding" ||
-    (recommendation.subject_type === "guided_onboarding_session" && recommendation.subject_id === sessionId) ||
-    metadataMatchesSession(recommendation.metadata, sessionId);
-}
-
-export function filterGuidedAnalysisRecommendations(
-  recommendations: AiRecommendationRecord[],
-  sessionId: string,
-): AiRecommendationRecord[] {
-  const onboardingScoped = recommendations.filter((recommendation) =>
-    isOnboardingScopedRecommendation(recommendation, sessionId),
-  );
-
-  if (onboardingScoped.length > 0) return onboardingScoped;
-  return recommendations.filter((recommendation) => recommendation.domain === "shop_boost");
 }
 
 async function loadGuidedAnalysisRecommendations(sessionId: string): Promise<AiRecommendationRecord[]> {
